Add client packet handler to avoid loading client classes on server and fix packets missing on server
